Plot lines with a color that depends on flags
Plot lines going from pointsStart to pointsEnd with a color that depends on flags. Each element in flag corresponds to a column in pointsStart and one in pointsEnd. If an element in flag is true, the corresponding line should be plotted in green, while if it is false, it should be plotted in red.