Plot points with a color
Plot points with a color that depends on flags. Each element in flag corresponds to a column in points (i.e., to a point). If an element in flag is true, the corresponding point should be plotted in green, while if it is false, it should be plotted in red.